Moved it all around and got things working
