+++ FinalEngine.Editor.Desktop/Views/Dialogs/Layout/ManageWindowLayoutsView.xaml.cs
@@ -5,7 +5,7 @@
 namespace FinalEngine.Editor.Desktop.Views.Dialogs.Layout;
 
 using FinalEngine.Editor.ViewModels.Dialogs.Layout;
-using FinalEngine.Editor.ViewModels.Interactions;
+using FinalEngine.Editor.ViewModels.Services.Interactions;
 using MahApps.Metro.Controls;
 
 /// <summary>
