+++ FinalEngine.Editor.Desktop/Views/Inspectors/PropertiesView.xaml
@@ -4,13 +4,12 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:inspectors="clr-namespace:FinalEngine.Editor.Desktop.Views.Inspectors"
-             xmlns:vm="clr-namespace:FinalEngine.Editor.ViewModels.Docking.Tools.Inspectors;assembly=FinalEngine.Editor.ViewModels"
-             xmlns:vme="clr-namespace:FinalEngine.Editor.ViewModels.Inspectors;assembly=FinalEngine.Editor.ViewModels"
+             xmlns:vm="clr-namespace:FinalEngine.Editor.ViewModels.Inspectors;assembly=FinalEngine.Editor.ViewModels"
              mc:Ignorable="d"
              d:DataContext="{d:DesignInstance Type=vm:PropertiesToolViewModel}">
 
     <UserControl.Resources>
-        <DataTemplate DataType="{x:Type vme:EntityInspectorViewModel}">
+        <DataTemplate DataType="{x:Type vm:EntityInspectorViewModel}">
             <inspectors:EntityInspectorView />
         </DataTemplate>
     </UserControl.Resources>
