+++ FinalEngine.Editor.Desktop/Views/Scenes/EntitySystemsView.xaml
@@ -3,7 +3,7 @@
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:vm="clr-namespace:FinalEngine.Editor.ViewModels.Docking.Tools.Scenes;assembly=FinalEngine.Editor.ViewModels"
+             xmlns:vm="clr-namespace:FinalEngine.Editor.ViewModels.Scenes;assembly=FinalEngine.Editor.ViewModels"
              mc:Ignorable="d"
              d:DataContext="{d:DesignInstance Type=vm:EntitySystemsToolViewModel}">
     <Grid>
