+++ FinalEngine.Editor.ViewModels/Inspectors/IConsoleToolViewModel.cs
@@ -2,7 +2,9 @@
 // Copyright (c) Software Antics. All rights reserved.
 // </copyright>
 
-namespace FinalEngine.Editor.ViewModels.Docking.Tools.Inspectors;
+namespace FinalEngine.Editor.ViewModels.Inspectors;
+
+using FinalEngine.Editor.ViewModels.Docking.Tools;
 
 /// <summary>
 /// Defines an interface that represents a model of a console tool view.
