+++ FinalEngine.Editor.ViewModels/Inspectors/IPropertiesToolViewModel.cs
@@ -2,9 +2,10 @@
 // Copyright (c) Software Antics. All rights reserved.
 // </copyright>
 
-namespace FinalEngine.Editor.ViewModels.Docking.Tools.Inspectors;
+namespace FinalEngine.Editor.ViewModels.Inspectors;
 
 using CommunityToolkit.Mvvm.ComponentModel;
+using FinalEngine.Editor.ViewModels.Docking.Tools;
 
 /// <summary>
 /// Defines an interface that represents a model of a properties tool view.
