+++ FinalEngine.Editor.ViewModels/Projects/IProjectExplorerToolViewModel.cs
@@ -2,12 +2,14 @@
 // Copyright (c) Software Antics. All rights reserved.
 // </copyright>
 
-namespace FinalEngine.Editor.ViewModels.Docking.Tools.Projects;
+namespace FinalEngine.Editor.ViewModels.Projects;
+
+using FinalEngine.Editor.ViewModels.Docking.Tools;
 
 /// <summary>
 /// Defines an interface that represents a model of a project explorer tool view.
 /// </summary>
-/// <seealso cref="FinalEngine.Editor.ViewModels.Docking.Tools.IToolViewModel" />
+/// <seealso cref="IToolViewModel" />
 public interface IProjectExplorerToolViewModel : IToolViewModel
 {
 }
