+++ FinalEngine.Editor.ViewModels/Scenes/IEntitySystemsToolViewModel.cs
@@ -2,7 +2,9 @@
 // Copyright (c) Software Antics. All rights reserved.
 // </copyright>
 
-namespace FinalEngine.Editor.ViewModels.Docking.Tools.Scenes;
+namespace FinalEngine.Editor.ViewModels.Scenes;
+
+using FinalEngine.Editor.ViewModels.Docking.Tools;
 
 /// <summary>
 /// Defines an interface that represents a model of an entity systems tool view.
