+++ FinalEngine.Editor.ViewModels/Services/Interactions/ICloseable.cs
@@ -2,7 +2,7 @@
 // Copyright (c) Software Antics. All rights reserved.
 // </copyright>
 
-namespace FinalEngine.Editor.ViewModels.Interactions;
+namespace FinalEngine.Editor.ViewModels.Services.Interactions;
 
 /// <summary>
 /// Defines an interface that represents an interaction to close a view.
