+++ FinalEngine.Tests/Editor/ViewModels/Docking/DockViewModelTests.cs
@@ -8,10 +8,9 @@ using System;
 using System.Linq;
 using FinalEngine.Editor.Common.Services.Factories;
 using FinalEngine.Editor.ViewModels.Docking;
-using FinalEngine.Editor.ViewModels.Docking.Panes.Scenes;
-using FinalEngine.Editor.ViewModels.Docking.Tools.Inspectors;
-using FinalEngine.Editor.ViewModels.Docking.Tools.Projects;
-using FinalEngine.Editor.ViewModels.Docking.Tools.Scenes;
+using FinalEngine.Editor.ViewModels.Inspectors;
+using FinalEngine.Editor.ViewModels.Projects;
+using FinalEngine.Editor.ViewModels.Scenes;
 using FinalEngine.Editor.ViewModels.Services.Layout;
 using Microsoft.Extensions.Logging;
 using Moq;
