+++ SharedAssemblyInfo.cs
@@ -9,8 +9,8 @@ using System.Reflection;
 [assembly: AssemblyCopyright("© 2023 Software Antics")]
 [assembly: AssemblyTrademark("Software Antics™")]
 [assembly: AssemblyCulture("")]
-[assembly: AssemblyVersion("2023.3.2932.0")]
-[assembly: AssemblyFileVersion("2023.3.2931.0")]
+[assembly: AssemblyVersion("2023.3.3028.0")]
+[assembly: AssemblyFileVersion("2023.3.3027.0")]
 #if DEBUG
 [assembly: AssemblyConfiguration("Debug")]
 #else
